requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
imgaug>=0.4.0

# Data Manipulation and Analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Explainability and Interpretability
shap>=0.42.0
lime>=0.2.0.1
captum>=0.6.0
grad-cam>=1.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Jupyter and Notebook Support
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Web Application Framework
streamlit>=1.28.0
dash>=2.14.0
flask>=2.3.0

# Data Loading and Processing
h5py>=3.9.0
pydicom>=2.4.0
nibabel>=5.1.0

# Model Serialization
joblib>=1.3.0
pickle5>=0.0.12

# Configuration and Logging
pyyaml>=6.0
configparser>=5.3.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development Tools
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
urllib3>=2.0.0
pathlib2>=2.3.7

# Medical Imaging Specific
SimpleITK>=2.2.0
pydicom>=2.4.0
medpy>=0.4.0

# Performance Monitoring
psutil>=5.9.0
GPUtil>=1.4.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Optional: GPU Support (uncomment if using CUDA)
# torch-cuda>=2.0.0
# cupy-cuda11x>=12.0.0  # Replace with appropriate CUDA version 